+++ src/commands/mod.rs
@@ -3,6 +3,7 @@ pub mod create;
 pub mod interactive;
 pub mod list;
 pub mod merge_pr_github;
+pub mod open_editor;
 
 pub mod pr_github;
 pub mod rm;