+++ src/lib.rs
@@ -1,6 +1,8 @@
 pub mod cli;
 mod commands;
+pub mod editor;
 mod repo;
+pub mod telemetry;
 
 pub use commands::create;
 pub use repo::Repo;