+++ tests/commands.rs
@@ -6,6 +6,8 @@ mod create;
 mod list;
 #[path = "commands/merge_pr_github.rs"]
 mod merge_pr_github;
+#[path = "commands/open_editor.rs"]
+mod open_editor;
 #[path = "commands/pr_github.rs"]
 mod pr_github;
 #[path = "commands/rm.rs"]